+++ migrations/0002_create_table_datapoint.py
@@ -1,0 +1,16 @@
+"""
+create table datapoint
+date created: 2019-02-14 22:01:50.586924
+"""
+
+
+def upgrade(migrator):
+    with migrator.create_table('datapoint') as table:
+        table.int('datapoint_id')
+        table.foreign_key('INT', 'metric_id', on_delete=None, on_update=None, references='metric.metric_id')
+        table.float('value')
+        table.int('timestamp')
+
+
+def downgrade(migrator):
+    migrator.drop_table('datapoint')